rename ultility index.js to messaging.js
